Animatable Version
Added function: SS_StarscapeManager.moveView(PVector). This simulates
parallax by moving the rendered positions of astronomical features in x
and y, but not in z. This is not suitable for use in realtime as it
takes far too long to render each frame.

Additional tweaks were made to stellar distribution and depth opacity
algorithms.